+++ src/components/Card/Card.tsx
@@ -1,9 +1,9 @@
 import { getFormattedValidityPeriodUnit } from "@utils/card";
 import masterCard from "@assets/Mastercard.png";
 import visa from "@assets/Visa.png";
-import type { CardNumberUnits } from "@components/CardNumberInputField";
-import type { ValidityPeriod } from "@components/CardValidityPeriodInputField";
-import SwitchCase from "@components/SwitchCase";
+import type { CardNumberUnits } from "@/components/CardNumberInputField/CardNumberInputField";
+import type { ValidityPeriod } from "@/components/CardValidityPeriodInputField/CardValidityPeriodInputField";
+import SwitchCase from "@components/common/SwitchCase";
 import styled from "@emotion/styled";
 import { detectCardBrand } from "@utils/card";
 import { COLOR_PALETTE } from "@/styles/colorPalette";
