+++ src/stories/CardCVCInputField.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 
-import CardCVCInputField from "../components/CardCVCInputField";
+import CardCVCInputField from "../components/CardCVCInputField/CardCVCInputField";
 
 const meta = {
   title: "CardCVCInputField",
